pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gen-ai-app"
version = "0.1.0"
description = "Minimal PDF indexing and question answering app with CLI and FastAPI entrypoints."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.2.0",
  "sentence-transformers>=3.0.0",
  "faiss-cpu>=1.8.0",
  "pypdf>=4.2.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "qdrant-client>=1.9.0",
  "langchain-openai>=0.1.0",
]

[tool.setuptools]
packages = ["app"]

[project.scripts]
genai-cli = "app.cli:main"
